backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
motor==3.3.1
pydantic>=2.6.4
email-validator>=2.2.0
python-multipart>=0.0.9
Pillow>=10.0.0
qrcode>=7.4.0
anthropic==0.40.0
httpx>=0.27.0
